backend/requirements.txt
bcrypt==4.2.1
blinker==1.8.2
click==8.1.7
coverage==7.6.1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
psycopg2==2.9.7
PyJWT==2.10.0
pytest==8.3.3
pytest-cov==5.0.0
python-dotenv==0.21.0
SQLAlchemy==2.0.21
typing_extensions==4.12.2
Werkzeug==3.0.4
